Add an IAM profile, role, and policy to the Nessus EC2 configuration
The policy allows access to SSM parameters that store the Nessus credentials
used by the local installation. The role and profile allow the instance to
access the resources described in this policy. This will allow an appropriate
job to access these SSM parameters at run-time instead of needing to hardcode
the values.